Added SimpleInjector test
