+++ build.sbt
@@ -4,8 +4,8 @@ organization := "me.frmr.lift"
 
 scalaVersion := "2.11.4"
 
-libraryDependencies += "net.liftweb" %% "lift-util" % "3.0-SNAPSHOT"
+libraryDependencies += "net.liftweb" %% "lift-util" % "3.0-M5-1"
 
-libraryDependencies += "net.liftweb" %% "lift-webkit" % "3.0-SNAPSHOT"
+libraryDependencies += "net.liftweb" %% "lift-webkit" % "3.0-M5-1"
 
 mainClass in Compile := Some("me.frmr.lift.PerformanceTester")
